April 21st Java Programs

+++ QT_231_232_CoreJava_Programs/src/basic_Programs/Addition2Num.java
@@ -5,8 +5,8 @@ public class Addition2Num
 
 	public static void main(String args[]) 
 	{
-		int num1=25;
-		int num2=35;
+		byte num1=25;
+		byte num2=35;
 		int sum=num1+num2;
 		System.out.println("Addition: "+sum); //60
